Update README.md
Goal of the Project
A tidy data set
A link to a Github repository with your script for performing the analysis
A code book that describes the variables, the data, and any transformations or work that you performed to clean up the data called CodeBook.md. You should also include a README.md in the repo with your scripts. This repo explains how all of the scripts work and how they are connected.
Analysis R Script